Docker setup
hi there,
We dockerized the laravel app, this way
we don't have to install anything in our local
machine (pending to update Readme.md).
We could move the ports to the .env in case
we need to run multiple docker containers (we
will), we now have phpunit only detail is that
in order to work with IntelliJ (AFAIK) we need
to give the ide it's own php interpreter and
that's why we have two php services.

We missed the mysql extensions that php needs to
work but are now added a `docker build` in your
terminal should do the work.

And finally we just added our small php.ini; we
already had it in the docker-compose.yml but it
was not being used.