Product service
The delete functionality is in place
now we're able to delete products.

Notes:
we're still not using groups in the
api.php file

We have a mixed in the request validations
I like to have control over certain things
and I keep the trade with the framework
having custom errors in ValidationRuleException
and letting Laravel handling the ones in rules.
The test cases (and clients) are the ones that
will see a difference between them and personally
I think making the all custom is better (but it's
a trade of having to do more work)